shop_system/requirements.txt
Django==2.0
django-allauth==0.36.0
django-bootstrap4==0.0.6

